Refactor namespaces used.
This changes the namespace for the gem from MultipartPost into
Multipart::Post to match the actual name of the gem.

This also extracts out of the global namespace the Parts module to live
instead under Multipart::Post::Parts
