+++ lib/multipart/post.rb
@@ -1,0 +1,4 @@
+module Multipart
+  module Post
+  end
+end
