+++ lib/multipart/post/version.rb
@@ -1,0 +1,5 @@
+module Multipart
+  module Post
+    VERSION = "2.1.1"
+  end
+end
