+++ spec/fixtures/multibyte.txt
